Apply newer AppArmor rules based on AppArmor parser capabilities
Adjust apparmor/gen.go to add Comment<Type>Rule strings to the data struct.
Then reference these in front of the corresponding rule in the baseTemplate.
Add various functions to see if the parser supports the rule type, and then
set data.Comment<Type>Rule to '#' for that rule if it does not. In this manner,
AppArmor policy should work with any combination of rule types.

Signed-off-by: Jamie Strandboge <jamie@ubuntu.com>